requirements.txt
uuid
pyautogui
opencv-python
pytesseract
Pillow
dotenv
screeninfo
numpy